requirements.txt
asn1crypto==0.24.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
firebase-admin==2.18.0
Flask==1.1.1
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.0
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.2
Werkzeug==0.15.5
win-inet-pton==1.1.0
wincertstore==0.2
